modified to a more general way
