+++ opencompass/datasets/__init__.py
@@ -20,6 +20,7 @@ from .cmnli import *  # noqa: F401, F403
 from .cmrc import *  # noqa: F401, F403
 from .commonsenseqa import *  # noqa: F401, F403
 from .copa import *  # noqa: F401, F403
+from .corev2 import *  # noqa: F401, F403
 from .crowspairs import *  # noqa: F401, F403
 from .csl import *  # noqa: F401, F403
 from .cvalues import *  # noqa: F401, F403
@@ -64,7 +65,7 @@ from .siqa import *  # noqa: F401, F403
 from .squad20 import SQuAD20Dataset, SQuAD20Evaluator  # noqa: F401, F403
 from .storycloze import *  # noqa: F401, F403
 from .strategyqa import *  # noqa: F401, F403
-from .subject_corev2 import *  # noqa: F401, F403
+from .subject import *  # noqa: F401, F403
 from .summedits import *  # noqa: F401, F403
 from .summscreen import *  # noqa: F401, F403
 from .TabMWP import *  # noqa: F401, F403
